Implemented reusable GameForm component
