+++ front/src/components/GameForm/style.scss
